Optimised size

+++ timeGloveNew.ino
@@ -1,16 +1,16 @@
 #include <Wire.h>
-#include "Kalman.h"
+#include "Kalman.h" // Kalman Filter Library
 
 const int mpu_addr = 0x68; // I2C address of the MPU-6050
-int light = 200;					 //задержка, свет включен, микросекунд
-int dark;									 //свет выключен, микросекунды
-int min_dark = 1;					 //миминмальная задержка темноты
-int max_dark = 50;				 //максимальная задержка темноты
+uint8_t light = 200;			 //задержка, свет включен, микросекунд
+uint8_t dark;							 //свет выключен, микросекунды
+uint8_t min_dark = 1;			 //миминмальная задержка темноты
+uint8_t max_dark = 50;		 //максимальная задержка темноты
 byte light_pin = 2;				 //сюда подключен свет
 byte potent_pin = 6;			 //аналоговый пин потенциометра
-int angle;
+int16_t angle;
 
-double accZ, accX, accY, gyroX, gyroZ;
+int16_t accZ, accX, accY, gyroX, gyroZ;
 double accZangle, accXangle;
 double kalAngleX, kalAngleZ;
 
@@ -28,23 +28,23 @@ void setup()
 	Wire.write(0); // wake up the mpu6050
 	Wire.endTransmission(true);
 	Serial.begin(9600);
-  pinMode(light_pin, OUTPUT);
-  pinMode(potent_pin, INPUT);
-  kalmanX.setAngle(180); // Set starting angle
-  kalmanZ.setAngle(180);
-  timer = micros();
+	pinMode(light_pin, OUTPUT);
+	pinMode(potent_pin, INPUT);
+	kalmanX.setAngle(180); // Set starting angle
+	kalmanZ.setAngle(180);
+	timer = micros();
 }
 void loop()
 {
 	measure();
- 
-		angle = kalAngleZ;																				 //вычисляем угол, сделав поправку на 250 градусов
-		dark = map(analogRead(potent_pin), 0, 1024, min_dark, max_dark); //расчет времени темноты как сумму угла с датчика угла и значения с потенциометра
-		digitalWrite(light_pin, 1);																			 // Включаем свет
-		delayMicroseconds(light);																				 // ждем
-		digitalWrite(light_pin, 0);																			 // выключаем
-		delay(dark);
-		delayMicroseconds(2000 + angle * 10); // ждем
+
+	angle = kalAngleZ;																							 //вычисляем угол, сделав поправку на 250 градусов
+	dark = map(analogRead(potent_pin), 0, 1024, min_dark, max_dark); //расчет времени темноты как сумму угла с датчика угла и значения с потенциометра
+	digitalWrite(light_pin, 1);																			 // Включаем свет
+	delayMicroseconds(light);																				 // ждем
+	digitalWrite(light_pin, 0);																			 // выключаем
+	delay(dark);
+	delayMicroseconds(2000 + angle * 10); // ждем
 }
 
 void measure()
@@ -64,29 +64,29 @@ void measure()
 	timer = micros();
 }
 
-double Gyro_X()
+int16_t Gyro_X()
 {
 	Wire.beginTransmission(mpu_addr);
 	Wire.write(0x43);
 	Wire.endTransmission(false);
 	Wire.requestFrom(mpu_addr, 2);
-	double GyX = Wire.read() << 8 | Wire.read();
+	int16_t GyX = Wire.read() << 8 | Wire.read();
 	Wire.endTransmission(true);
 	return GyX;
 }
 
-double Gyro_Z()
+int16_t Gyro_Z()
 {
 	Wire.beginTransmission(mpu_addr);
 	Wire.write(0x47);
 	Wire.endTransmission(false);
 	Wire.requestFrom(mpu_addr, 2);
-	double GyZ = Wire.read() << 8 | Wire.read();
+	int16_t GyZ = Wire.read() << 8 | Wire.read();
 	Wire.endTransmission(true);
 	return GyZ;
 }
 
-double Acc_Y()
+int16_t Acc_Y()
 {
 	Wire.beginTransmission(mpu_addr);
 	Wire.write(0x3D);
@@ -97,24 +97,24 @@ double Acc_Y()
 	return AccY;
 }
 
-double Acc_X()
+int16_t Acc_X()
 {
 	Wire.beginTransmission(mpu_addr);
 	Wire.write(0x3B);
 	Wire.endTransmission(false);
 	Wire.requestFrom(mpu_addr, 2);
-	double AccX = Wire.read() << 8 | Wire.read();
+	int16_t AccX = Wire.read() << 8 | Wire.read();
 	Wire.endTransmission(true);
 	return AccX;
 }
 
-double Acc_Z()
+int16_t Acc_Z()
 {
 	Wire.beginTransmission(mpu_addr);
 	Wire.write(0x3F);
 	Wire.endTransmission(false);
 	Wire.requestFrom(mpu_addr, 2);
-	double AccZ = Wire.read() << 8 | Wire.read();
+	int16_t AccZ = Wire.read() << 8 | Wire.read();
 	Wire.endTransmission(true);
 	return AccZ;
 }
